complete refactor for additional codebase clarity
