+++ automata/cli/scripts/run_code_embedding.py
@@ -7,12 +7,10 @@ from automata.config.base import ConfigCategory
 from automata.core.base.database.vector import JSONEmbeddingVectorDatabase
 from automata.core.coding.py.module_loader import py_module_loader
 from automata.core.llm.providers.openai import OpenAIEmbeddingProvider
-from automata.core.memory_store.symbol_code_embedding import (
-    SymbolCodeEmbeddingBuilder,
-    SymbolCodeEmbeddingHandler,
-)
+from automata.core.memory_store.symbol_code_embedding import SymbolCodeEmbeddingHandler
 from automata.core.symbol.graph import SymbolGraph
 from automata.core.symbol.symbol_utils import get_rankable_symbols
+from automata.core.symbol_embedding.embedding_builders import SymbolCodeEmbeddingBuilder
 from automata.core.utils import get_config_fpath
 
 logger = logging.getLogger(__name__)
