+++ automata/cli/scripts/run_doc_embedding_l3.py
@@ -14,19 +14,17 @@ from automata.core.llm.providers.openai import (
     OpenAIChatCompletionProvider,
     OpenAIEmbeddingProvider,
 )
-from automata.core.memory_store.symbol_code_embedding import (
-    SymbolCodeEmbeddingBuilder,
-    SymbolCodeEmbeddingHandler,
-)
-from automata.core.memory_store.symbol_doc_embedding import (
-    SymbolDocEmbeddingBuilder,
-    SymbolDocEmbeddingHandler,
-)
+from automata.core.memory_store.symbol_code_embedding import SymbolCodeEmbeddingHandler
+from automata.core.memory_store.symbol_doc_embedding import SymbolDocEmbeddingHandler
 from automata.core.symbol.base import SymbolDescriptor
 from automata.core.symbol.graph import SymbolGraph
 from automata.core.symbol.search.rank import SymbolRankConfig
 from automata.core.symbol.search.symbol_search import SymbolSearch
 from automata.core.symbol.symbol_utils import get_rankable_symbols
+from automata.core.symbol_embedding.embedding_builders import (
+    SymbolCodeEmbeddingBuilder,
+    SymbolDocEmbeddingBuilder,
+)
 from automata.core.symbol_embedding.similarity import SymbolSimilarityCalculator
 from automata.core.utils import get_config_fpath
