+++ automata/config/symbol/symbol_doc_embedding_l2.json
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:e60d740a1f17cbfc55b46516b9d203547a305d53acdc2074a89b2f9e39c95de1
-size 5155686
+oid sha256:102f68cb8b8b5dc8ca810e86cb83e14dbf62abcd70c6da24b569be7af05b614f
+size 5157496
