+++ automata/config/symbol/symbol_doc_embedding_l3.json
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:0877e565ceb063fd4bddcfc942a45d6051625f1bd31f462ae11d2e8917c6abba
-size 5776327
+oid sha256:0ab289ef890683e90465a0c38a7f63561f2ae5e0814e06ab8dad731f56870ff1
+size 5778177
