+++ automata/core/agent/tool/tool_utils.py
@@ -20,13 +20,11 @@ from automata.core.llm.providers.openai import (
     OpenAIEmbeddingProvider,
 )
 from automata.core.memory_store.symbol_code_embedding import SymbolCodeEmbeddingHandler
-from automata.core.memory_store.symbol_doc_embedding import (
-    SymbolDocEmbeddingBuilder,
-    SymbolDocEmbeddingHandler,
-)
+from automata.core.memory_store.symbol_doc_embedding import SymbolDocEmbeddingHandler
 from automata.core.symbol.graph import SymbolGraph
 from automata.core.symbol.search.rank import SymbolRank, SymbolRankConfig
 from automata.core.symbol.search.symbol_search import SymbolSearch
+from automata.core.symbol_embedding.embedding_builders import SymbolDocEmbeddingBuilder
 from automata.core.symbol_embedding.similarity import SymbolSimilarityCalculator
 from automata.core.utils import get_config_fpath
