+++ automata/docs/core/base/database/symbol_database_provider.rst
@@ -10,7 +10,7 @@ Related Symbols
 ---------------
 
 -  ``automata.core.symbol.base.Symbol``
--  ``automata.core.symbol.base.SymbolEmbedding``
+-  ``automata.core.symbol_embedding.base.SymbolEmbedding``
 -  ``automata.core.memory_store.symbol_code_embedding.SymbolCodeEmbeddingHandler``
 -  ``automata.core.base.database.vector.JSONEmbeddingVectorDatabase``
 -  ``automata.tests.utils.factories.symbol_search_live``
