+++ automata/docs/core/base/database/vector_database_provider.rst
@@ -26,7 +26,7 @@ Related Symbols
 -  ``automata.core.memory_store.symbol_code_embedding.SymbolCodeEmbeddingHandler``
 -  ``automata.core.base.database.provider.SymbolDatabaseProvider``
 -  ``automata.core.symbol.base.Symbol``
--  ``automata.core.symbol.base.SymbolEmbedding``
+-  ``automata.core.symbol_embedding.base.SymbolEmbedding``
 
 Example
 -------
