+++ automata/docs/core/database/vector_database_provider.rst
@@ -22,7 +22,7 @@ Related Symbols
 -  ``automata.core.database.provider.SymbolDatabaseProvider``
 -  ``automata.core.database.vector.JSONEmbeddingVectorDatabase``
 -  ``automata.core.symbol.base.Symbol``
--  ``automata.core.symbol.base.SymbolEmbedding``
+-  ``automata.core.symbol_embedding.base.SymbolEmbedding``
 -  ``automata.core.memory_store.embedding_types.SymbolEmbeddingHandler``
 -  ``automata.core.memory_store.symbol_code_embedding.SymbolCodeEmbeddingHandler``
