+++ automata/docs/core/llm/embedding_norm_type.rst
@@ -13,9 +13,9 @@ Related Symbols
 -  ``automata.tests.unit.test_symbol_embedding.test_add_new_embedding``
 -  ``automata.core.symbol_embedding.similarity.SymbolSimilarity``
 -  ``automata.tests.conftest.mock_embedding``
--  ``automata.core.symbol.base.SymbolEmbedding``
+-  ``automata.core.symbol_embedding.base.SymbolEmbedding``
 -  ``automata.tests.unit.test_symbol_similarity.test_get_nearest_symbols_for_query``
--  ``automata.core.symbol.base.SymbolCodeEmbedding``
+-  ``automata.core.symbol_embedding.base.SymbolCodeEmbedding``
 -  ``automata.tests.unit.test_symbol_embedding.test_update_embedding``
 
 Example
